v1.2 released
Adds five new modes, including DIVE, which I've wanted to add to the Power Compendium for a while now.

I'm aware of a potential bug where the game temporarily freezes upon moving after the tab went inactive, but I have no idea if the bug still exists or how to test for it... so hopefully it's gone.